Centralize theme/scopedcssbaseline code for embedded
