+++ packages/app-core/package.json
@@ -44,13 +44,13 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.16.3",
+    "@mui/icons-material": "^5.0.0",
     "@mui/material": "^5.10.17",
     "copy-to-clipboard": "^3.3.1",
     "react-error-boundary": "^4.0.3",
     "shortid": "^2.2.15"
   },
   "peerDependencies": {
-    "@mui/icons-material": "^5.0.0",
     "mobx": "^6.0.0",
     "mobx-react": "^7.0.0",
     "mobx-state-tree": "^5.0.0",
