+++ packages/embedded-core/package.json
@@ -44,6 +44,7 @@
   "dependencies": {
     "@babel/runtime": "^7.16.3",
     "@mui/material": "^5.10.17",
+    "@mui/icons-material": "^5.0.0",
     "@jbrowse/product-core": "^2.5.0",
     "copy-to-clipboard": "^3.3.1",
     "react-error-boundary": "^4.0.3",
