+++ packages/embedded-core/src/index.ts
@@ -1,2 +1,2 @@
 export { default as ModalWidget } from './ui/ModalWidget'
-export { default as EmbeddedViewContainer } from './ui/ViewContainer'
+export { default as EmbeddedViewContainer } from './ui/EmbeddedViewContainer'
