+++ packages/product-core/package.json
@@ -42,6 +42,7 @@
     "useSrc": "node ../../scripts/useSrc.js"
   },
   "dependencies": {
+    "@mui/icons-material": "^5.0.0",
     "@babel/runtime": "^7.16.3",
     "@mui/material": "^5.10.17",
     "copy-to-clipboard": "^3.3.1",
