+++ products/jbrowse-react-circular-genome-view/src/JBrowseCircularGenomeView/JBrowseCircularGenomeView.tsx
@@ -3,28 +3,16 @@ import { observer } from 'mobx-react'
 import { getEnv } from '@jbrowse/core/util'
 import { readConfObject } from '@jbrowse/core/configuration'
 import { createJBrowseTheme } from '@jbrowse/core/ui'
-import { ThemeProvider } from '@mui/material/styles'
-import { ScopedCssBaseline } from '@mui/material'
-import { makeStyles } from 'tss-react/mui'
 import { EmbeddedViewContainer, ModalWidget } from '@jbrowse/embedded-core'
 
 // locals
 import { ViewModel } from '../createModel/createModel'
-
-const useStyles = makeStyles()({
-  // avoid parent styles getting into this div
-  // https://css-tricks.com/almanac/properties/a/all/
-  avoidParentStyle: {
-    all: 'initial',
-  },
-})
 
 const JBrowseCircularGenomeView = observer(function ({
   viewState,
 }: {
   viewState: ViewModel
 }) {
-  const { classes } = useStyles()
   const { session } = viewState
   const { view } = session
   const { pluginManager } = getEnv(session)
@@ -38,18 +26,14 @@ const JBrowseCircularGenomeView = observer(function ({
   )
 
   return (
-    <ThemeProvider theme={theme}>
-      <div className={classes.avoidParentStyle}>
-        <ScopedCssBaseline>
-          <EmbeddedViewContainer key={`view-${view.id}`} view={view}>
-            <Suspense fallback={<div>Loading...</div>}>
-              <ReactComponent model={view} session={session} />
-            </Suspense>
-          </EmbeddedViewContainer>
-          <ModalWidget session={session} />
-        </ScopedCssBaseline>
-      </div>
-    </ThemeProvider>
+    <>
+      <EmbeddedViewContainer key={`view-${view.id}`} view={view} theme={theme}>
+        <Suspense fallback={<div>Loading...</div>}>
+          <ReactComponent model={view} session={session} />
+        </Suspense>
+      </EmbeddedViewContainer>
+      <ModalWidget session={session} />
+    </>
   )
 })
